flowctl: add `raw bundle` for dumping resolved inline catalogs
Refactor out the implementation from the `draft author` subcommand.
